requirements.txt.txt
langchain
openai
unstructured
unstructured[local-inference]
detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2
poppler-utils
tiktoken
pinecone-client
